updated readme, moar improvements sb

+++ soundboard.sh
@@ -119,35 +119,44 @@ clean_dir() {
 }
 
 create_sounds_from_file() {
+	read -p "Slice from local configurations? (y/n): " slice_local_choice
 
-	clean_dir
+	if [ ${slice_local_choice:-"n"} == "y" ]; then
+		file_to_slice="temp_file_to_slice.mp3"	
+	else
+		clean_dir
 
-	[[ -f "ranges_values.txt" ]] && rm -rf ranges_values.txt
+		[[ -f "ranges_values.txt" ]] && rm -rf ranges_values.txt
         
-	read -p "Enter filename with path: " file_to_slice
+		read -p "Enter filename with path: " file_to_slice
 	    
-	polished_range_in="default"
+		polished_range_in="default"
 	        
-	while  [ "$polished_range_in" != "exit" ]; do
-		read -p "Enter slicing info [starting(seconds),timetoadd(seconds),author,title] es: 30,70,fabi,siamai. Enter to finish: " range_in
-		polished_range_in=${range_in:-"exit"}
-		[[ $polished_range_in != "exit" ]] && echo "$polished_range_in" >> ranges_values.txt
-	done
+		while  [ "$polished_range_in" != "exit" ]; do
+			read -p "Enter slicing info [starting(seconds),timetoadd(seconds),author,title] es: 30,70,fabi,siamai. Enter to finish: " range_in
+			polished_range_in=${range_in:-"exit"}
+			[[ $polished_range_in != "exit" ]] && echo "$polished_range_in" >> ranges_values.txt
+		done
 
-	check_file=$(basename "$file_to_slice" | rev | cut -f 1 -d '.' | rev)
+		check_file=$(basename "$file_to_slice" | rev | cut -f 1 -d '.' | rev)
 
-	[[ ! "$check_file" =~ ^(mp3|wav)$ ]] && echo "Converting to mp3..." && ffmpeg -i "$file_to_slice" temp_file_to_slice.mp3 >/dev/null 2>&1 && file_to_slice="temp_file_to_slice.mp3"
+		[[ ! "$check_file" =~ ^(mp3|wav)$ ]] && echo "Converting to audio..." && ffmpeg -i "$file_to_slice" temp_file_to_slice.mp3 >/dev/null 2>&1 && file_to_slice="temp_file_to_slice.mp3"
+	fi
+
 
 	exec 3<ranges_values.txt
 	while IFS= read -u 3 line; do
-		echo "Reading file $line"
+#		echo "Reading file $line"
 		starting=$(echo "$line" | cut -d ',' -f 1)
 		time_add=$(echo "$line" | cut -d ',' -f 2)
 		artist_slice=$(echo "$line" | cut -d ',' -f 3)
 		title_slice=$(echo "$line" | cut -d ',' -f 4)
-		ffmpeg -ss "$starting" -t "$time_add" -i "$file_to_slice" -metadata artist="${artist_slice:-"Unknown"}" -metadata title="$title_slice" -acodec copy $SOUNDS_DIR/$(($(ls $SOUNDS_DIR | wc -l) + 1)).mp3 >/dev/null 2>&1
-		echo "Sound created for -> $artist_slice: $title_slice"
+		counter=1 #counter stays 1 fix, even tho a sounds without a title is bad
+		ffmpeg -ss "$starting" -t "$time_add" -i "$file_to_slice" -metadata artist="${artist_slice:-"Unknown"}" -metadata title="${title_slice:-"unknown_track_$counter"}" -acodec copy $SOUNDS_DIR/$(($(ls $SOUNDS_DIR | wc -l) + 1)).mp3 >/dev/null 2>&1
+		(( counter++ ))
+#		echo "Sound created for -> $artist_slice: $title_slice"
 	done
+	echo "Sounds created"
 }
 
 
@@ -201,6 +210,6 @@ while [[ $WHATDO != 'exit' ]]; do
 	esac
 done
 
-clean_dir
+#clean_dir
 
 #sudo telegram-send --file sudo telegram-send "message"
